Removed GigaChannelCommunity
