+++ src/tribler/core/components/popularity/community/sync_strategy.py
